@@ -1,38 +1,34 @@
-class Mutations::ChangeFavorites < Mutations::BaseMutation
-  description "お気に入り一括変更"
+# frozen_string_literal: true
 
-  argument :artist_ids, [TTID], required: false, description: "お気に入り変更したいアーティストID"
-  argument :album_ids, [TTID], required: false, description: "お気に入り変更したいアルバムID"
-  argument :track_ids, [TTID], required: false, description: "お気に入り変更したいトラックID"
-  argument :playlist_ids, [TTID], required: false, description: "お気に入り変更したいプレイリストID"
-  argument :radio_ids, [TTID], required: false, description: "お気に入り変更したいラジオID"
-  argument :favorite, Boolean, required: true, description: "true の場合は一括でお気に入り登録をする。false 場合は一括で解除する。"
+module Mutations
+  class ChangeFavorites < ::Mutations::BaseMutation
+    description 'お気に入り一括変更'
 
-  field :current_user, CurrentUserType, null: true, description: "更新されたカレントユーザー"
-  field :error, String, null: true
+    argument :artist_ids, [::String], required: false, description: 'お気に入り変更したいアーティストID'
+    argument :album_ids, [::String], required: false, description: 'お気に入り変更したいアルバムID'
+    argument :track_ids, [::String], required: false, description: 'お気に入り変更したいトラックID'
+    argument :playlist_ids, [::String], required: false, description: 'お気に入り変更したいプレイリストID'
+    argument :favorite,
+             ::GraphQL::Types::Boolean,
+             required: true,
+             description: 'true の場合は一括でお気に入り登録をする。false 場合は一括で解除する。'
 
-  def mutate(artist_ids: [], album_ids: [], track_ids: [], playlist_ids:[], radio_ids: [], favorite:)
-    begin
-      artists = ::Artist.where(id: artist_ids)
-      albums  = ::Album.where(id: album_ids)
-      tracks  = ::Track.where(id: track_ids)
-      playlists = ::Playlist.where(id: playlist_ids)
-      radios = ::Radio.where(id: radio_ids)
+    field :current_user, ::Types::Objects::CurrentUserObject, null: true, description: '更新されたカレントユーザー'
+
+    def mutate(favorite:, artist_ids: [], album_ids: [], track_ids: [], playlist_ids: [])
+      artists = ::Artist.where(id: artist_ids).to_a
+      albums  = ::Album.where(id: album_ids).to_a
+      tracks  = ::Track.where(id: track_ids).to_a
+      playlists = ::Playlist.where(id: playlist_ids).to_a
 
       if favorite
-        ::Favorite.register(artists + albums + tracks + playlists + radios, context[:current_info][:user])
+        ::Favorite.register(artists + albums + tracks + playlists, context[:current_info][:user])
       else
-        ::Favorite.unregister(artists + albums + tracks + playlists + radios, context[:current_info][:user])
+        ::Favorite.unregister(artists + albums + tracks + playlists, context[:current_info][:user])
       end
 
       {
-        current_user: context[:current_info][:user].reload,
-        error: nil,
-      }
-    rescue => error
-      {
-        current_user: nil,
-        error: error.message,
+        current_user: context[:current_info][:user].reload
       }
     end
   end
