@@ -1,21 +1,16 @@
-class Mutations::ClearCache < Mutations::BaseMutation
-  description "検索結果キャッシュをリセットする"
+# frozen_string_literal: true
 
-  field :results, [String], null: true
-  field :error, String, null: true
+module Mutations
+  class ClearCache < ::Mutations::BaseMutation
+    description '検索結果キャッシュをリセットする'
 
-  def mutate
-    begin
-      size = File::Stat.new(Rails.root.join("tmp", "cache")).size.to_s(:human_size, locale: :en)
+    field :result, [::String], null: true
+
+    def mutate
+      size = ::File::Stat.new(::Rails.root.join('tmp', 'cache')).size.to_s
 
       {
-        results: [size, *Rails.cache.clear],
-        error: nil,
-      }
-    rescue => error
-      {
-        results: nil,
-        error: error.message,
+        result: [size, *::Rails.cache.clear]
       }
     end
   end
