@@ -1,13 +1,19 @@
-class Mutations::IgnoreArtists < Mutations::BaseMutation
-  description "PENDING状態のアーティスト全てをIGNOREにする。よく考えてから実行すること。"
+# frozen_string_literal: true
 
-  field :artists, [ArtistType], null: true, description: "IGNOREされたアーティスト"
+module Mutations
+  class IgnoreArtists < ::Mutations::BaseMutation
+    description 'PENDING状態のアーティスト全てをIGNOREにする。よく考えてから実行すること。'
 
-  def mutate
-    artists = Artist.all_pending_to_ignore
-    Rails.cache.clear
-    {
-      artists: artists,
-    }
+    field :result, ::GraphQL::Types::Boolean, null: true
+
+    def mutate
+      ::Artist.all_pending_to_ignore
+
+      ::Rails.cache.clear
+
+      {
+        result: true
+      }
+    end
   end
 end
