@@ -1,30 +1,29 @@
-class Mutations::Login < Mutations::BaseMutation
-  description "ログイン"
+# frozen_string_literal: true
 
-  argument :username, String, required: true
-  argument :current_password, String, required: true
+module Mutations
+  class Login < ::Mutations::BaseMutation
+    description 'ログイン'
 
-  field :current_user, Types::Objects::CurrentUserType, null: true
+    argument :username, ::String, required: true
+    argument :current_password, ::String, required: true
 
-  def use_recaptcha?; true end
+    field :current_user, ::Types::Objects::CurrentUserObject, null: true
 
-  def mutate(username:, current_password:)
-    user = User.find_by(username: username)
+    def use_recaptcha?() = true
 
-    unless user&.authenticate(current_password)
-      raise GraphQL::ExecutionError.new(
-        "ユーザーIDまたはパスワードに誤りがあります" ,
-        extensions: { code: "FAILED_AUTHENTICATION" }
-      )
+    def mutate(username:, current_password:)
+      user = ::User.find_by(username: username)
+
+      unless user && user.authenticate(current_password)
+        raise(::GraphQL::ExecutionError.new('ユーザーIDまたはパスワードに誤りがあります', extensions: { code: 'FAILED_AUTHENTICATION' }))
+      end
+
+      context[:current_info][:user] = user
+      context[:current_info][:session] = user.create_session
+
+      {
+        current_user: user
+      }
     end
-
-    # cookie 更新
-    # 複数デバイスを許可しておく
-    context[:current_info][:user] = user
-    context[:current_info][:session] = user.sessions.create!
-
-    {
-      current_user: user,
-    }
   end
 end
