@@ -1,25 +1,18 @@
-class Mutations::Logout < Mutations::BaseMutation
-  description "ログアウト"
+# frozen_string_literal: true
 
-  field :current_user, Types::Objects::CurrentUserType, null: true
-  field :error, String, null: true
+module Mutations
+  class Logout < ::Mutations::BaseMutation
+    description 'ログアウト'
 
-  def mutate
-    begin
-      user = User.create_user_and_session!
+    field :current_user, ::Types::Objects::CurrentUserObject, null: true
 
-      # cookie 更新
+    def mutate
+      user = ::User.create_user_and_session!
       context[:current_info][:user]    = user
       context[:current_info][:session] = user.sessions.first
 
       {
-        current_user: user,
-        error: nil,
-      }
-    rescue => error
-      {
-        current_user: nil,
-        error: "ログアウトに失敗しました",
+        current_user: user
       }
     end
   end
