@@ -1,30 +1,33 @@
-class Mutations::UpsertPlaylist < Mutations::BaseMutation
-  description "プレイリストを作成更新する"
+# frozen_string_literal: true
 
-  argument :playlist_id, TTID, required: false, description: "IDがない場合は作成"
-  argument :name, String, required: true, description: "タイトル"
-  argument :description, String, required: false, description: "説明"
-  argument :public_type, PlaylistPublicTypeEnum, required: true, description: "公開種別"
-  argument :track_ids, [TTID], required: false, description: "トラックID"
+module Mutations
+  class UpsertPlaylist < ::Mutations::BaseMutation
+    description 'プレイリストを作成更新する'
 
-  field :playlist, PlaylistType, null: true, description: "作成更新されたプレイリスト"
-  field :error, String, null: true
+    argument :playlist_id, ::String, required: false, description: 'IDがない場合は作成'
+    argument :name, ::String, required: true, description: 'タイトル'
+    argument :description, ::String, required: false, description: '説明'
+    argument :public_type, ::Types::Enums::PlaylistPublicTypeEnum, required: true, description: '公開種別'
+    argument :track_ids, [::String], required: false, description: 'トラックID'
 
-  def mutate(playlist_id: nil, name:, description: "", public_type:, track_ids: [])
-    Playlist.validate_author(playlist_id, context[:current_info][:user].id) if playlist_id
+    field :playlist, ::Types::Objects::PlaylistObject, null: true, description: '作成更新されたプレイリスト'
 
-    playlist =
-      Playlist.upsert(
-        id: playlist_id,
-        user_id: context[:current_info][:user].id,
-        name: name,
-        description: description,
-        public_type: public_type,
-        track_ids: track_ids
-      )
+    def mutate(name:, public_type:, playlist_id: nil, description: '', track_ids: [])
+      ::Playlist.validate_author!(playlist_id, context[:current_info][:user].id) if playlist_id
 
-    {
-      playlist: playlist,
-    }
+      playlist =
+        ::Playlist.create_or_update(
+          id: playlist_id,
+          user_id: context[:current_info][:user].id,
+          name: name,
+          description: description,
+          public_type: public_type,
+          track_ids: track_ids
+        )
+
+      {
+        playlist: playlist
+      }
+    end
   end
 end
