@@ -1,6 +1,8 @@
+# frozen_string_literal: true
+
 module Types
   module Enums
-    class BaseEnum < GraphQL::Schema::Enum
+    class BaseEnum < ::GraphQL::Schema::Enum
     end
   end
 end
