@@ -1,8 +1,10 @@
+# frozen_string_literal: true
+
 module Types
   module InputObjects
-    class CursorInputObject < BaseInputObject
-      argument :offset, Integer, required: false, default_value: 0
-      argument :limit,  Integer, required: false, default_value: 30
+    class CursorInputObject < ::Types::InputObjects::BaseInputObject
+      argument :offset, ::Integer, required: false, default_value: 0
+      argument :limit,  ::Integer, required: false, default_value: 50
     end
   end
 end
