@@ -1,9 +1,8 @@
+# frozen_string_literal: true
+
 module Types
   module Objects
-    class BaseObject < GraphQL::Schema::Object
-      include Types::Objects
-      include Types::Scalars
-      include Types::Enums
+    class BaseObject < ::GraphQL::Schema::Object
     end
   end
 end
