@@ -1,80 +1,78 @@
-class Artist < ApplicationRecord
-  table_id :art
+# frozen_string_literal: true
 
-  include Artists::Status
-  include Artists::Mix
+class Artist < ::ApplicationRecord
+  def table_id() = 'art'
 
   has_many :artist_has_albums, dependent: :destroy
   has_many :albums, through: :artist_has_albums
   has_many :artist_has_tracks, dependent: :destroy
   has_many :tracks, through: :artist_has_tracks
   has_many :apple_music_artists, dependent: :destroy
-
   has_many :favorites, as: :favorable, dependent: :destroy
-
-  scope :include_album_services, -> { eager_load(albums: :apple_music_and_itunes_album) }
-  scope :include_albums, -> { eager_load(:albums) }
-  scope :include_tracks, -> { eager_load(:tracks) }
-  scope :include_services, -> { eager_load(:apple_music_artists) }
-  scope :services, -> { include_services.map(&:service) }
-  scope :names, -> { services.map(&:name) }
 
   enum status: { pending: 0, active: 1, ignore: 2 }
 
-  FORCE_IGNORE_NAMES = %w[V.A. ヴァリアス・アーティスト]
-
   class << self
-    def create_by_name(name)
-      name = to_name(name)
-      find_or_create_by!(name: name)
-      AppleMusicArtist.create_by_name(name)
-      where(name: name)
+    def cache?(conditions:)
+      cache = true
+      cache = false if conditions.key?(:favorite)
+      cache
     end
 
-    def to_name(name)
-      name = name.to_s.
-             # 英数を全角から半角へ変換
-             tr('０-９ａ-ｚＡ-Ｚ','0-9a-zA-Z').
-             # カタカナを半角から全角へ変換
-             sub(/[\uFF61-\uFF9F]+/) { |str| str.unicode_normalize(:nfkc) }.
-             # 半角全角スペースを削除して英語名の場合は名前ごとに区切ってキャメルケースにして半角スペースを追加する
-             split(/\p{blank}/).map {|n| n.camelize }.join(" ").gsub(/([^A-Za-z\s])\s([^A-Za-z\s])/, '\1\2')
-      raise StandardError, "空文字のアーティスト名になっている" unless name.present?
-      name
+    def generate_relation(conditions:, context:)
+      # @type var album_ids: ::Array[::String]
+      # @type var artist_ids: ::Array[::String]
+      # @type var name: ::String
+
+      conditions = { status: [:active] }.merge(conditions)
+      relation = ::Artist.includes(albums: :apple_music_album)
+
+      if conditions.key?(:name)
+        name = conditions.delete(:name)
+        relation = relation.where('artists.name like :name', name: "%#{name}%")
+      end
+
+      if conditions.key?(:album_ids)
+        albums = ::Album.includes(:artists).includes(:tracks).where(id: conditions.delete(:album_ids))
+        artist_ids = albums.map { |a| a.artists.ids }
+                           .flatten
+        artist_ids += albums.map { |a| a.tracks.includes(:artists).map { |t| t.artists.ids } }
+                            .flatten
+        relation = relation.where(id: artist_ids)
+      end
+
+      if conditions.key?(:track_ids)
+        relation = relation.includes(:tracks).where(tracks: { id: conditions.delete(:track_ids) })
+      end
+
+      if conditions.delete(:favorite)
+        relation = relation.joins(:favorites).where(favorites: { user_id: context[:current_info][:user].id })
+      end
+
+      relation.where(conditions)
     end
 
     def all_pending_to_ignore
-      where(status: :pending).map do |artist|
-        artist.ignore!
-        artist
-      end
-    end
-
-    def all_create_albums
-      active_artists = where(status: :active)
-      rate = active_artists.size / 100.to_f
-      active_artists.map.with_index do |artist, index|
-        Rails.logger.info("---------------- #{artist.name} #{(index / rate).round}% ----------------")
-        artist.create_albums
-      end
-      Rails.cache.clear
+      where(status: :pending).update_all(status: :ignore)
     end
   end
 
-  def create_albums
-    apple_music_artists.each {|am_art| am_art.create_albums }
-    self
+  def artwork_l
+    return ::Artwork.new unless (album = service_album)
+
+    album.artwork_l
   end
 
-  def service
-    @service ||= (apple_music_artists.to_a).first
+  def artwork_m
+    return ::Artwork.new unless (album = service_album)
+
+    album.artwork_m
   end
 
-  def to_path
-    "/#{id}"
-  end
+  def service_album
+    return unless (album = albums.first)
+    return unless (apple_music_album = album.apple_music_album)
 
-  def to_url
-    "#{ENV['PRODUCTION_APP_URL']}/artists#{to_path}"
+    apple_music_album
   end
 end
