@@ -1,5 +1,7 @@
-class ArtistHasAlbum < ApplicationRecord
-  table_id :ahb
+# frozen_string_literal: true
+
+class ArtistHasAlbum < ::ApplicationRecord
+  def table_id() = 'ahb'
 
   belongs_to :artist
   belongs_to :album
