@@ -1,6 +1,8 @@
+# frozen_string_literal: true
+
 class Artwork
-  include ActiveModel::Model
-  include ActiveModel::Attributes
+  include ::ActiveModel::Model
+  include ::ActiveModel::Attributes
 
   attribute :url,    :string
   attribute :width,  :integer
