@@ -1,8 +1,8 @@
-class PlaylistItem < ApplicationRecord
-  table_id :psi
+# frozen_string_literal: true
+
+class PlaylistItem < ::ApplicationRecord
+  def table_id() = 'psi'
 
   belongs_to :playlist
   belongs_to :track
-
-  scope :include_tracks, -> { eager_load(track: [:apple_music_and_itunes_tracks]) }
 end
