@@ -1,20 +1,16 @@
-class Role < ApplicationRecord
-  table_id :rol
+# frozen_string_literal: true
+
+class Role < ::ApplicationRecord
+  def table_id() = 'rol'
 
   has_many :users
   has_many :allowed_actions, dependent: :destroy
 
   class << self
-    def admin_role
-      Role.find_by!(name: "admin")
-    end
+    def admin() = find_by!(name: 'admin')
 
-    def login_role
-      Role.find_by!(name: "login")
-    end
+    def login() = find_by!(name: 'login')
 
-    def default_role
-      Role.find_by!(name: "default")
-    end
+    def default() = find_by!(name: 'default')
   end
 end
