@@ -1,2 +1,4 @@
-# クッキーシリアライズ方式を json にする
-Rails.application.config.action_dispatch.cookies_serializer = :json
+# frozen_string_literal: true
+
+# クッキーシリアライズ方式は JSON
+::Rails.application.config.action_dispatch.cookies_serializer = :json
